Implement "spread" feature
Code will replace first spread character ('|' ATM) in Bottom and
Side text with as many spaces as possible.  This result in two
parts of the string (left/right, top/bottom) to be spread as far
as possible from each other.  With that you can align some parts
ot the formatted string to one side of the screen and second part
to another part.

In practice this can be used to display seconds in bottom right
corner and still have date in bottom left corner.  Similar in the
Side text we can now have battery charge percentage on the bottom.