v4.1.0 finalize implementation of spread character
Change spread character from '|' to ',' as it is easier to type
on mobile phone.  Also the spread operation makes no sense on
round displays so everything is available only for rectangle
Pebbles.  I added some documentation in Clay config with one
new example for Bottom and Side texts.  There is a slight issue
with examples as I have the same liest of preesets for all
platforms, and as I said earlier the sperad operator makes no
sense on Pebble Round.  But the comma character is no supported
and it's displayed as empty space anyway, so new examples also
works for Round.